requirements.txt
pytest
requests
pytest-html
allure-pytest
pyyaml
